requirements.txt
openai-whisper
watchdog
